pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bombercat-integrator"
version = "0.1.0"
description = "BomberCat IoT Device Integration Platform"
authors = [{name = "BomberCat Team", email = "team@bombercat.com"}]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "esptool>=4.6.2",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "pyserial>=3.5",
    "numpy>=1.24.0",
    "flet==0.28.*",
    "awsiotsdk>=1.11.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "awscrt>=0.19.0",
    "awsiot>=0.2.0",
    "tenacity>=8.2.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.setuptools.package-dir]
"" = "modules"

[tool.setuptools.packages.find]
where = ["modules"]
include = ["bombercat_*"]
namespaces = true

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
bombercat-integrator = "api.main:app"